Introduces kerberos provider replacement
