+++ src/onegov/user/auth/__init__.py
@@ -1,3 +1,4 @@
 from onegov.user.auth.core import Auth
+from onegov.user.auth.kerberos import KerberosClient
 
-__all__ = ('Auth', )
+__all__ = ('Auth', 'KerberosClient')
