+++ tests/onegov/user/test_integration.py
@@ -4,9 +4,10 @@ import morepath
 import pytest
 
 from onegov.core import Framework
+from onegov.core.security import Public
 from onegov.core.utils import scan_morepath_modules
-from onegov.user import Auth, UserApp
-from onegov.user.auth.provider import provider_by_name
+from onegov.user import UserApp
+from onegov.user.auth import KerberosClient
 from tempfile import NamedTemporaryFile
 from unittest.mock import patch, DEFAULT
 from webtest import TestApp as Client
@@ -46,10 +47,17 @@ def app(request, postgres_dsn, temporary_path, redis_url, keytab):
     class App(Framework, UserApp):
         pass
 
-    # Auth needs to be linkable
-    @App.path(model=Auth, path='/auth')
-    def get_auth(request):
-        return None
+    @App.path(model=KerberosClient, path='/kerberos-client')
+    def get_kerberos_client(request):
+        return KerberosClient(
+            keytab=keytab,
+            hostname='ogc.example.org',
+            service='HTTP'
+        )
+
+    @App.view(model=KerberosClient, permission=Public)
+    def view_kerberos_client(self, request):
+        return self.authenticated_username(request)
 
     scan_morepath_modules(App)
     morepath.commit(App)
@@ -58,13 +66,6 @@ def app(request, postgres_dsn, temporary_path, redis_url, keytab):
     app.configure_application(
         dsn=postgres_dsn,
         redis_url=redis_url,
-        authentication_providers={
-            'kerberos': {
-                'keytab': keytab,
-                'hostname': 'ogc.example.org',
-                'service': 'HTTP'
-            }
-        }
     )
 
     app.namespace = 'apps'
@@ -78,13 +79,11 @@ def client(app):
     yield Client(app)
 
 
-def test_kerberos_auth(client, app):
-    provider = provider_by_name(app.providers, 'kerberos')
-    assert provider is not None
+def test_kerberos_client(client, app, keytab):
 
     def auth(headers=None):
         return client.get(
-            url='/auth/provider/kerberos',
+            url='/kerberos-client',
             expect_errors=True,
             headers=headers
         )
@@ -121,12 +120,11 @@ def test_kerberos_auth(client, app):
         assert r.status_code == 401
         assert r.headers['WWW-Authenticate'] == 'Negotiate'
 
-        # simulate a successful exchange, with no associated user
+        # simulate a successful exchange
         methods['authGSSServerInit'].return_value = 1, None
         methods['authGSSServerStep'].return_value = 1
         methods['authGSSServerResponse'].return_value = 'barfoo'
         methods['authGSSServerUserName'].return_value = 'foo@EXAMPLE.ORG'
         r = auth({'Authorization': 'Negotiate foobar'})
 
-        assert r.status_code == 302
-        assert r.location.endswith('/')
+        assert r.text == 'foo@EXAMPLE.ORG'
